Backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
google-genai>=1.0.0
jinja2==3.1.3
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
pillow>=10.2.0